requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain_core
python-dotenv
pydantic
flask